requirements.txt
django
isort
black
djangorestframework
djangorestframework-simplejwt
django-filter